stop using email to open account and change float to integer

+++ apps/bank_logic/priv/repo/migrations/20180822234743_add_transactions_table.exs
@@ -4,7 +4,7 @@ defmodule BankLogic.Repo.Migrations.AddTransactionsTable do
   def change do
     create table(:transactions) do
       add :operation, :integer
-      add :amount, :float
+      add :amount, :integer
       add :source_id, references(:accounts, on_delete: :delete_all), null: false
       add :destination_id, references(:accounts, on_delete: :delete_all), null: true
